Overhaul BigInteger implementation.
This provides a faster code path for #valueOf(long) and some other
operations, cleans up constructors and also removes the JavaScript
specific native code.

PiperOrigin-RevId: 361980135

+++ user/super/com/google/gwt/emul/java/math/BigDecimal.java
@@ -2572,7 +2572,7 @@ public class BigDecimal extends Number implements Comparable<BigDecimal>,
 
   private BigInteger getUnscaledValue() {
     if (intVal == null) {
-      intVal = BigInteger.valueOf(smallValue);
+      intVal = BigInteger.valueOf((long) smallValue);
     }
     return intVal;
   }
